Choose a variant from the form to select all fitted, fitted original, fitted alternate or not fitted components on a PCB
Choose a variant from the form to select all fitted, fitted original,
fitted alternate or not fitted components on a PCB.
This script creates a filter expression to select components that are
fitted in a variant. The filter expression is placed in the PCB Filter
panel and applied. This can then be used to create variant component
classes in the PCB panel.