Update Distribute script to suppress misleading error message
If script is run with only 2 tracks selected, it would tell you to select more than 3 tracks. Script runs normally on 3 or more tracks, and will run the by-value options on 2 or more tracks. This change suppresses the message if 2 tracks are selected, but disables the options that require 3 or more tracks. In addition, since any message now means that the script will fail, do not bring up the script dialog after the message.