Update distribute script (#25)
* Patch 3 (#1)

* Update Fillet.pas to deselect ineligible objects

Script will fail if non-track objects are included in the selection, so during startup, deselect such items.

* Update Fillet.dfm form

Minor quality-of-life improvement to clean up form and make it not resizable.

* Update Distribute script to suppress misleading error message

If script is run with only 2 tracks selected, it would tell you to select more than 3 tracks. Script runs normally on 3 or more tracks, and will run the by-value options on 2 or more tracks. This change suppresses the message if 2 tracks are selected, but disables the options that require 3 or more tracks. In addition, since any message now means that the script will fail, do not bring up the script dialog after the message.

* Updated Distribute form default setting and Escape

It seems more often that evenly distributing clearances between traces is needed. Also, I think it's generally helpful if the "Escape" key cancels dialogs so that is enabled now.

* Revert "Updated Distribute form default setting and Escape"

This reverts commit 5edc4b033052b5c1fd7b8b1d814111a6caedc248.

* Revert "Update Distribute script to suppress misleading error message"

This reverts commit ccc6e7a4563cd8e3fe3fc2103db32131b9edc630.

* Update Distribute script to suppress misleading error message

If script is run with only 2 tracks selected, it would tell you to select more than 3 tracks. Script runs normally on 3 or more tracks, and will run the by-value options on 2 or more tracks. This change suppresses the message if 2 tracks are selected, but disables the options that require 3 or more tracks. In addition, since any message now means that the script will fail, do not bring up the script dialog after the message.

* Updated Distribute form default setting and Escape

It seems more often that evenly distributing clearances between traces is needed. Also, I think it's generally helpful if the "Escape" key cancels dialogs so that is enabled now.

* Revert "Patch 3 (#1)"

This reverts commit 45ac768ab3335c6847c1d171b2bcf8e94610d7ac.

* Create README.md

added documentation