Update AutoPlaceSilkscreen.pas to fix autoposition justification
Fixed autoposition justification, the bane of my existence. It turns out I convolved rotation with the autoposition behavior based on aspect ratio of the text's bounding box. 

In short, rotation determines justification, but aspect ratio determines whether "TopLeft" is placed on Top or Left first. 

If text's bounding box is taller than it is wide (aka vertical), TopLeft means placed on the left side, aligned with the top edge. If bounding box is not vertical, TopLeft means placed on the top side, aligned with the left edge.